+++ packages/node/src/methods/index.ts
@@ -3,13 +3,11 @@ import {
   getInstalledAppInstancesController,
   getProposedAppInstancesController
 } from "./app-instance-operations";
-import { installAppInstanceController } from "./install/controller";
-import {
-  createMultisigController,
-  getAllChannelAddressesController
-} from "./multisig-operations";
-import { proposeInstallVirtualAppInstanceController } from "./propose-install-virtual/controller";
+import installAppInstanceController from "./install/controller";
+import proposeInstallVirtualAppInstanceController from "./propose-install-virtual/controller";
 import { proposeInstallAppInstanceController } from "./propose-install/controller";
+import createMultisigController from "./state-channel/create/controller";
+import getAllChannelAddressesController from "./state-channel/get-all/controller";
 
 export {
   installAppInstanceController,
