+++ packages/node/src/methods/install/controller.ts
@@ -11,7 +11,7 @@ import { install } from "./app-instance";
  * sending an approved ack to the proposer.
  * @param params
  */
-export async function installAppInstanceController(
+export default async function installAppInstanceController(
   this: RequestHandler,
   params: Node.InstallParams
 ): Promise<Node.InstallResult> {
